refactor(runners/models): harmonize pilot/training file split
- Split openvla.py -> openvla.py (inference) + openvla_train.py (training)
- Rename gr00t.py -> gr00t_train.py (training-only; pilot lives in evals/)
- Update re-exports in runners/models/__init__.py and runners/__init__.py

Closes #98
