+++ src/odyssey/runners/models/__init__.py
@@ -1,18 +1,28 @@
 """Model loaders — training runners and inference models."""
 
+from odyssey.runners.models.gr00t_train import (
+    GR00TRunner,
+    build_gr00t_argv,
+    parse_gr00t_line,
+)
 from odyssey.runners.models.openvla import (
+    make_openvla_policy,
+)
+from odyssey.runners.models.openvla_train import (
     OpenVLARunner,
     build_openvla_argv,
-    make_openvla_policy,
     parse_openvla_line,
 )
 
 __all__ = [
+    "GR00TRunner",
     "GemmaVLMGenerator",
     "OpenVLARunner",
     "VLARuntime",
+    "build_gr00t_argv",
     "build_openvla_argv",
     "make_openvla_policy",
+    "parse_gr00t_line",
     "parse_openvla_line",
 ]
