+++ src/odyssey/runners/models/pi05_train.py
@@ -63,7 +63,7 @@ from odyssey.runners.base import (
 )
 
 # Shared flatten helper (dotted keys for nested dicts).
-from odyssey.runners.models.openvla import _flatten_config
+from odyssey.runners.models.openvla_train import _flatten_config
 from odyssey.runners.subprocess import (
     TrainingProcessSpec,
     output_path,
@@ -231,11 +231,8 @@ def _lerobot_env_for_dataset(task: TrainingTask) -> dict[str, str]:
     if task.dataset is None:
         return {}
     ref = task.dataset.ref
-    if task.dataset.source == DatasetSource.LOCAL and os.path.isabs(ref):
-        parent = os.path.dirname(os.path.normpath(ref))
-        # Only HF_LEROBOT_HOME — do NOT also set the legacy LEROBOT_HOME. Recent
-        # lerobot HARD-FAILS at import (raises ValueError) if the deprecated
-        # LEROBOT_HOME var is present, which killed compute_norm_stats.
+    if task.dataset.source == DatasetSource.LOCAL and (os.path.isabs(ref) or ref.startswith("/")):
+        parent = Path(ref).parent.as_posix()
         return {"HF_LEROBOT_HOME": parent}
     return {}
